CmakeList cleaned up. bugfix equations -> equation.
Introduced grouping by cmake variable name, so that in ccmake, where everything is alphabetical, the variables remain grouped.

Also if an additional cmake variable is set in an if , so that it depends on another cmake variable, it is now unset by default.
This way if for example PARABOLIC is switched off again, the LIFTING variable also disappears...

Some options moved to Advanced ,in ccmake, they are then found with t for toggle